refactor: fix localization name in all files
