+++ packages/core/src/auth/i18n/en.json
@@ -20,7 +20,7 @@
   "User_Language": "Language",
   "User_Theme": "Theme",
   "User_NoAppUser": "The user could not be fetched. Please verify that you have the necessary permissions or contact your administrator.",
-  "Auth_SliceAction_FetchLanguages": "fetch languages",
+  "Auth_SliceAction_FetchLocalization": "fetch localizations",
   "Auth_SliceAction_FetchCompanies": "fetch companies",
   "Auth_SliceAction_FetchActiveUser": "fetch active user",
   "Auth_SliceAction_UpdateActiveUser": "update active user"
